Refactor useBackend() hook and WebSocketHandler.java.

+++ src/main/java/org/igor/onlinegames/xogame/dto/XoGameStateDto.java
@@ -13,6 +13,6 @@ public class XoGameStateDto implements XoGameDto {
     private XoGamePhase phase;
     private List<List<XoCellDto>> field;
     private List<XoPlayerDto> players;
-    private Long playerIdToMove;
-    private Long winnerId;
+    private Integer playerIdToMove;
+    private Integer winnerId;
 }
